pyproject.toml
[project]
name = "PromptLab"
version = "0.1.0"
description = "Captioning to Prompt and Text to Prompt. Tuned for Z-image, Flux.2 Klein and Qwen-Image."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "google-genai>=1.57.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.4",
    "openai>=1.30.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "requests>=2.32.3",
    "uvicorn>=0.30.0",
]
